feat(vs2022_third): C4018 fix signed/unsigned mismatch in NanoSpec — 7 files (batch 5)
Cast int loop variables to DWORD/unsigned int in comparisons:
- NanoSpec/StressLineSectionSettingDlg.cpp: 33 instances (i,j→DWORD)
- NanoSpec/PointConfirmationDlg.cpp: 1 instance (m_iNumScans→DWORD)
- NanoSpec/PointSetupTeaching2PointDlg.cpp: 1 instance (m_iNumScans→DWORD)
- NanoSpec/PointSetupTeachingDlg.cpp: 2 instances (m_iNumScans→DWORD)
- NanoSpec/SrVideoCalibrationSettingDlg.cpp: 2 instances (i→unsigned int)
- NanoSpec/StressElasticModulusListDlg.cpp: 2 instances (i→unsigned int)

Total: 41 C4018 warnings fixed.
No behavior changes, only explicit casts to resolve type mismatch.

+++ SRC/NanoSpec/PointConfirmationDlg.cpp
@@ -344,7 +344,7 @@ BOOL CPointConfirmationDlg::CheckInputCurrPoint()
 	GetDlgItemText(IDC_POINT_NO, sCurrPoint);
 	dwCurrPoint = strtol(sCurrPoint, NULL, 10);
 
-	if ((dwCurrPoint >= 1) && (dwCurrPoint < m_iNumScans + 1)){
+	if ((dwCurrPoint >= 1) && (dwCurrPoint < (DWORD)(m_iNumScans + 1))){
 		m_iCurrPoint = dwCurrPoint;
 	} else {
 		// Err
